+++ frontend/src/Pages/Home/CommentsModal.tsx
@@ -1,6 +1,6 @@
 import axios from 'axios';
 import React, { useEffect, useState } from 'react';
-import CommentsVotingModal from './CommentsVotingModal';
+import CommentModal from './CommentModal';
 
 interface Comment {
     commentId: string;
@@ -21,17 +21,6 @@ const CommentsModal: React.FC<ModalProps> = ({ isOpenState, threadId }) => {
     const [comments, setComments] = useState<Comment[]>([]);
     const [commentsWithNames, setCommentsWithNames] = useState<Comment[]>([]);
     const [newComment, setNewComment] = useState<string>("");
-
-    const [isOpenUpvote, setIsOpenUpvote] = useState<boolean>(false);
-    const [isOpenDownvote, setIsOpenDownvote] = useState<boolean>(false);
-
-    const handleOpenUpvoteModal = () => {
-        setIsOpenUpvote(true);
-    };
-
-    const handleOpenDownvoteModal = () => {
-        setIsOpenDownvote(true);
-    };
 
     useEffect(() => {
         const fetchCommentUserNames = async () => {
@@ -118,56 +107,6 @@ const CommentsModal: React.FC<ModalProps> = ({ isOpenState, threadId }) => {
         }
     };
 
-    const handleUpvote = async (commentId : string) => {
-        const token = localStorage.getItem('token');
-        const upvoteData = {
-            upvoter : token, // Set the authorId from JWT
-            threadId,
-            commentId
-        };
-        
-        try {
-            const response = await axios.post(`${import.meta.env.VITE_SERVER_URL}/threads/upvotecomment`, upvoteData,
-                {
-                    headers: {
-                        'Content-Type': 'application/json',
-                        Authorization: `Bearer ${token}`,
-                    },
-                }
-            );
-            const { updatedComment } = response.data;
-
-            setComments(updatedComment);  
-        } catch (err: any) {
-            console.log("error");
-        }
-    };
-
-    const handleDownvote = async (commentId : string) => {
-        const token = localStorage.getItem('token');
-        const downvoteData = {
-            downvoter : token, // Set the authorId from JWT
-            threadId,
-            commentId
-        };
-        
-        try {
-            const response = await axios.post(`${import.meta.env.VITE_SERVER_URL}/threads/downvotecomment`, downvoteData,
-                {
-                    headers: {
-                        'Content-Type': 'application/json',
-                        Authorization: `Bearer ${token}`,
-                    },
-                }
-            );
-            const { updatedComment } = response.data;
-
-            setComments(updatedComment);  
-        } catch (err: any) {
-            console.log("error");
-        }
-    };
-
     if (!isOpen) return null;
 
     return (
@@ -188,56 +127,7 @@ const CommentsModal: React.FC<ModalProps> = ({ isOpenState, threadId }) => {
                 <h2 className="text-xl font-semibold mb-4">Comments</h2>
                 <div className="flex-1 overflow-y-auto mb-4">
                     {commentsWithNames.map((comment) => (
-                        <div
-                            key={comment.commentId}
-                            className="flex flex-col bg-gray-100 transition p-3 rounded-lg shadow-sm mb-2"
-                        >
-                            <div className="flex justify-between items-center">
-                                <span className="text-gray-800 font-medium">{comment.userName}</span>
-                            </div>
-                            <p className="text-gray-700 mt-2">{comment.content}</p>
-                            <div className="flex items-center justify-between p-1">
-                                <div className="flex items-center space-x-4">
-                                    <div>
-                                        <button
-                                            onClick={() => handleUpvote(comment.commentId)}
-                                            className={`py-1 px-3 rounded-md border ${
-                                                false
-                                                    ? "bg-blue-500 text-white"
-                                                    : "bg-gray-200 text-gray-700 hover:bg-blue-500 hover:text-white"
-                                            }`}
-                                        >
-                                            ▲
-                                        </button>
-                                        <button onClick={handleOpenUpvoteModal} className="py-1 px-3 rounded-md text-gray-700 hover:text-blue-500">
-                                            {comment.upvotes.length}
-                                        </button>
-                                    </div>
-                                    <div>
-                                        <button
-                                            onClick={() => handleDownvote(comment.commentId)}
-                                            className={`py-1 px-3 rounded-md border ${
-                                                false
-                                                    ? "bg-red-500 text-white"
-                                                    : "bg-gray-200 text-gray-700 hover:bg-red-500 hover:text-white"
-                                            }`}
-                                        >
-                                            ▼
-                                        </button>
-                                        <button onClick={handleOpenDownvoteModal} className="py-1 px-3 rounded-md text-gray-700 hover:text-blue-500">
-                                            {comment.downvotes.length}
-                                        </button>
-                                    </div>
-                                </div>
-                                {isOpenUpvote && 
-                                    <CommentsVotingModal voteType="Upvotes" votes={comment.upvotes} isOpenState={[isOpenUpvote, setIsOpenUpvote]}></CommentsVotingModal>
-                                }
-
-                                {isOpenDownvote && 
-                                    <CommentsVotingModal voteType="Downvotes" votes={comment.downvotes} isOpenState={[isOpenDownvote, setIsOpenDownvote]}></CommentsVotingModal>
-                                }
-                            </div>
-                        </div>
+                        <CommentModal key={comment.commentId} comment={comment} threadId={threadId}></CommentModal>
                     ))}
                     {commentsWithNames.length === 0 && (
                         <p className="text-gray-600 text-center">No comments available.</p>
